Finished frontend

+++ Tourplanner.BusinessLayer/CalculateAttributes/ICalculateAttributes.cs
@@ -6,10 +6,10 @@ namespace Tourplanner.BusinessLayer
     {
         public bool CalculateChildFriendly(IEnumerable<Log> logs, double distance);
         public PopularityEnum CalculatePopularity(IEnumerable<Log> logs);
-        public float AverageRatingCalc(List<Log> logs);
+        public int AverageRatingCalc(List<Log> logs);
         public float AverageTimeCalc(List<Log> logs);
         public string CalcTimeFormated(int time);
-        public float AverageDifficultyCalc(List<Log> logs);
+        public int AverageDifficultyCalc(List<Log> logs);
         public string GetErrorString(IEnumerable<string?> errorMessages);
     }
 }
